20230119 #2
필터 수정

+++ src/main/java/com/kantar/vo/BaseVO.java
@@ -1,7 +1,5 @@
 package com.kantar.vo;
 
-import java.math.BigInteger;
-
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
@@ -11,7 +9,7 @@ import lombok.Setter;
 @RequiredArgsConstructor
 public class BaseVO{
     private Integer idx;
-    private BigInteger idx_user;
+    private Integer idx_user;
     private String create_dt;
     private String update_dt;
     private Integer currentPage;
